+++ examples/blog/src/pages/search-index.ts
@@ -1,21 +1,21 @@
-
-import { createIndex, indexToObject, LinearIndex, StaticSeekError } from "staticseek";
+import type { PageProps } from "cirrojs";
 import { markdownToText } from "cirrojs/server";
-import { posts } from "../lib/content";
+import { createIndex, indexToObject, LinearIndex, StaticSeekError } from "staticseek";
+import type { content } from "../content";
 
-export function genearteSearchIndex(): string {
-    const target = posts.map((md) => {
+export function generateSearchIndex(props: PageProps<typeof content>): string {
+    const target = props.content.posts.map((md) => {
         const content = markdownToText(md.content);
         return {
             ...md,
-            content
+            content,
         };
     });
 
     const index = createIndex(LinearIndex, target, {
         key_fields: ["slug", "title", "date", "tags"],
-        search_targets: ["title", "description", "content"]
-    })
+        search_targets: ["title", "description", "content"],
+    });
 
     if (index instanceof StaticSeekError) throw index;
     return JSON.stringify(indexToObject(index));